Revert "Regen historical states for `new-state-mgmt` compatibility (#5261)"
This reverts commit df9a534826037ddac8dcaac0b1b470ce9fa8ecd4.
