+++ beacon-chain/blockchain/service.go
@@ -471,10 +471,6 @@ func (s *Service) pruneGarbageState(ctx context.Context, slot uint64) error {
 		return err
 	}
 
-	if err := s.beaconDB.SaveLastArchivedIndex(ctx, 0); err != nil {
-		return err
-	}
-
 	return nil
 }
