+++ beacon-chain/db/kv/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "attestations.go",
         "backup.go",
         "blocks.go",
-        "check_historical_state.go",
+        "check_state.go",
         "checkpoint.go",
         "deposit_contract.go",
         "encoding.go",
@@ -16,7 +16,6 @@ go_library(
         "kv.go",
         "operations.go",
         "powchain.go",
-        "regen_historical_states.go",
         "schema.go",
         "slashings.go",
         "state.go",
@@ -29,7 +28,6 @@ go_library(
     deps = [
         "//beacon-chain/cache:go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/core/state:go_default_library",
         "//beacon-chain/db/filters:go_default_library",
         "//beacon-chain/db/iface:go_default_library",
         "//beacon-chain/state:go_default_library",
@@ -37,7 +35,6 @@ go_library(
         "//proto/beacon/db:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/bytesutil:go_default_library",
-        "//shared/cmd:go_default_library",
         "//shared/featureconfig:go_default_library",
         "//shared/params:go_default_library",
         "//shared/sliceutil:go_default_library",
